# docs/research/mvp/final/implementation/api/openapi.yaml
openapi: "3.1.0"
info:
  title: HelixTerminator API
  version: "1.0.0"
  description: |
    HelixTerminator — Next-Generation Enterprise SSH Client Platform.
    Dual-module product family: Module A (Secure Terminal Platform) + Module B (Zero-Trust Connection Broker).
  contact:
    name: HelixTerminator Engineering
    email: api@helixterminator.io
    url: https://helixterminator.io
  license:
    name: Proprietary
    identifier: Proprietary
servers:
  - url: https://api.helixterminator.io/api/v1
    description: Production
  - url: https://staging-api.helixterminator.io/api/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /hosts:
    get:
      summary: List hosts
      operationId: hostList
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 25 } }
        - { name: status, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200':
          description: Host list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Create host
      operationId: hostCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostCreateRequest'
      responses:
        '201':
          description: Host created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
  /hosts/{hostId}:
    get:
      summary: Get host
      operationId: hostGet
      parameters:
        - { name: hostId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Update host
      operationId: hostUpdate
      parameters:
        - { name: hostId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostUpdateRequest'
      responses:
        '200':
          description: Updated host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
    delete:
      summary: Delete host
      operationId: hostDelete
      parameters:
        - { name: hostId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Host deleted }
...
  /sessions:
    get:
      summary: List sessions
      operationId: sessionList
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 25 } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200':
          description: Session list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Create session
      operationId: sessionCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreateRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
  /sessions/{sessionId}:
    get:
      summary: Get session
      operationId: sessionGet
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
    delete:
      summary: Terminate session
      operationId: sessionTerminate
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Session terminated }
...
  /workspaces:
    get:
      summary: List workspaces
      operationId: workspaceList
      responses:
        '200':
          description: Workspace list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Create workspace
      operationId: workspaceCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceCreateRequest'
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
  /workspaces/{workspaceId}:
    get:
      summary: Get workspace
      operationId: workspaceGet
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
    put:
      summary: Update workspace
      operationId: workspaceUpdate
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceUpdateRequest'
      responses:
        '200':
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
    delete:
      summary: Delete workspace
      operationId: workspaceDelete
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Workspace deleted }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: EdDSA (Ed25519) signed JWT
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    Pagination:
      type: object
      properties:
        cursor_next: { type: string, nullable: true }
        cursor_prev: { type: string, nullable: true }
        has_next: { type: boolean }
        has_prev: { type: boolean }
        limit: { type: integer }
        total_count: { type: integer, nullable: true }
    PaginatedResponse:
      type: object
      properties:
        data: { type: array, items: { type: object } }
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Auth
...
    Host:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        hostname: { type: string }
        port: { type: integer, minimum: 1, maximum: 65535, default: 22 }
        username: { type: string }
        auth_method: { type: string, enum: [key, password, certificate, interactive, agent, pgp] }
        os: { type: string }
        os_version: { type: string }
        tags: { type: array, items: { type: string } }
        status: { type: string, enum: [active, inactive, unreachable, archived] }
        last_connected_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    HostCreateRequest:
      type: object
      required: [name, hostname]
      properties:
        name: { type: string }
        hostname: { type: string }
        port: { type: integer, default: 22 }
        username: { type: string }
        auth_method: { type: string, default: key }
        key_id: { type: string, format: uuid }
        os: { type: string }
        tags: { type: array, items: { type: string } }
        vault_id: { type: string, format: uuid }
        group_id: { type: string, format: uuid }
    HostUpdateRequest:
      type: object
      properties:
        name: { type: string }
        hostname: { type: string }
        port: { type: integer }
        username: { type: string }
        auth_method: { type: string }
        key_id: { type: string, format: uuid }
        os: { type: string }
        tags: { type: array, items: { type: string } }
        status: { type: string }
...
    Session:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        host_id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        status: { type: string, enum: [connecting, connected, disconnected, error, terminated] }
        terminal_cols: { type: integer, default: 80 }
        terminal_rows: { type: integer, default: 24 }
        recording_enabled: { type: boolean }
        collab_enabled: { type: boolean }
        started_at: { type: string, format: date-time }
        connected_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time }
        duration_seconds: { type: integer }
        bytes_sent: { type: integer }
        bytes_received: { type: integer }
    SessionCreateRequest:
      type: object
      required: [host_id]
      properties:
        host_id: { type: string, format: uuid }
        key_id: { type: string, format: uuid }
        auth_method: { type: string }
        terminal_cols: { type: integer, default: 80 }
        terminal_rows: { type: integer, default: 24 }
        recording_enabled: { type: boolean, default: false }
        collab_enabled: { type: boolean, default: false }

    # SFTP
...
    Workspace:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        layout: { type: object }
        is_template: { type: boolean }
        pinned: { type: boolean }
        auto_connect: { type: boolean }
        last_opened_at: { type: string, format: date-time }
        open_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WorkspaceCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        layout: { type: object }
        auto_connect: { type: boolean, default: true }
    WorkspaceUpdateRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        layout: { type: object }
        pinned: { type: boolean }
        auto_connect: { type: boolean }

    # Organization
...
  responses:
...
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            type: object
            properties:
              type: { type: string }
              title: { type: string }
              status: { type: integer }
              detail: { type: string }
...